feat: use struct pointers, break up handler methods
