+++ internal/host-builder/log.go
@@ -1,0 +1,35 @@
+package hostbuilder
+
+import (
+	"context"
+
+	"github.com/Cloud-RAMP/wasm-sandbox/internal/logging"
+	wasmevents "github.com/Cloud-RAMP/wasm-sandbox/pkg/wasm-events"
+	"github.com/tetratelabs/wazero/api"
+)
+
+func logHandler(handlerMap *wasmevents.HandlerMap) any {
+	return func(ctx context.Context, mod api.Module, strPtr uint32, strLen uint32) {
+		mem := mod.Memory()
+		if mem == nil {
+			return
+		}
+
+		bytes, ok := mem.Read(strPtr, strLen)
+		if !ok {
+			return
+		}
+		info := string(bytes)
+
+		event, err := getWASMEvent(ctx, wasmevents.LOG, info)
+		if event == nil {
+			logging.Logger.Errorf("Failed to create WASM event in handler %s: %v", wasmevents.LOG.String(), err)
+			return
+		}
+
+		_, err = handlerMap.CallHandler(event)
+		if err != nil {
+			logging.Logger.Errorf("Failed to call handler in %s: %v", wasmevents.LOG.String(), err)
+		}
+	}
+}
